LibraryScreen refactored to separate files
