+++ app/src/main/java/com/ovidium/comoriod/views/library/AuthorsCarousel.kt
@@ -1,0 +1,83 @@
+package com.ovidium.comoriod.views.library
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import com.ovidium.comoriod.views.DataItem
+
+@Composable
+fun AuthorsCarousel(
+    name: String,
+    dataItems: List<DataItem>?,
+    estimatedSize: Int,
+    isLoading: Boolean
+) {
+    val itemMinWidth = 180
+    val marginSize = 12
+    var screenWidth = LocalConfiguration.current.screenWidthDp
+    val itemsByRow = screenWidth / itemMinWidth
+
+    fun calculateItemSize(screenWidth: Int): Int {
+        return (screenWidth - (itemsByRow + 1) * marginSize) / itemsByRow
+    }
+
+    var itemSize = calculateItemSize(screenWidth)
+    val itemCnt = dataItems?.size ?: estimatedSize
+    if (itemCnt > itemsByRow) {
+        screenWidth -= (itemSize * 0.3).toInt()
+        itemSize = calculateItemSize(screenWidth)
+    }
+
+    Column(
+        verticalArrangement = Arrangement.spacedBy(marginSize.dp),
+        modifier = Modifier
+            .fillMaxHeight()
+            .padding(marginSize.dp)
+    ) {
+        Text(
+            name,
+            style = MaterialTheme.typography.h5,
+            fontWeight = FontWeight.Bold
+        )
+
+        LazyRow(horizontalArrangement = Arrangement.spacedBy(marginSize.dp)) {
+            if (isLoading)
+                repeat(estimatedSize) {
+                    item() {
+                        AuthorCard(
+                            title = "",
+                            isLoading = isLoading,
+                            colors = emptyList(),
+                            itemSize = itemSize,
+                            marginSize = marginSize
+                        )
+                    }
+                }
+            else {
+                dataItems?.let {
+                    items(dataItems) { dataItem ->
+                        AuthorCard(
+                            dataItem.title,
+                            isLoading,
+                            dataItem.imageId,
+                            dataItem.gradient,
+                            itemSize,
+                            marginSize
+                        )
+                    }
+
+                }
+            }
+        }
+    }
+}
