+++ app/src/main/java/com/ovidium/comoriod/views/library/RecentlyAddedBooksRow.kt
@@ -1,0 +1,31 @@
+package com.ovidium.comoriod.views.library
+
+import androidx.compose.runtime.Composable
+import com.ovidium.comoriod.data.recentlyaddedbooks.RecentlyAddedBooksResponse
+import com.ovidium.comoriod.data.recentlyaddedbooks.RecentlyAddedBooksResponseItem
+import com.ovidium.comoriod.mappings.getDrawableByAuthor
+import com.ovidium.comoriod.utils.getVolumeCoverGradient
+import com.ovidium.comoriod.views.DataItem
+
+@Composable
+fun RecentlyAddedBooksRow(response: RecentlyAddedBooksResponse?, isLoading: Boolean, isDark: Boolean) {
+    fun getAuthor(item: RecentlyAddedBooksResponseItem) : String? {
+        return if (item.authors.isEmpty()) null else item.authors[0]
+    }
+
+    val items = response?.map { item ->
+        val author = getAuthor(item)
+
+        DataItem(
+            item.name,
+            author,
+            item.volume,
+            author?.let { getDrawableByAuthor(author) },
+            getVolumeCoverGradient(item.volume, isDark),
+        )
+    }
+
+    val show = isLoading || (items != null && items.isNotEmpty())
+    if (show)
+        HorizontalCarousel("Adăugate recent", items, 2, isLoading)
+}
